+++ CHANGELOG.md
@@ -19,7 +19,7 @@ TODO: Provide proper migration guide
 * removed alias `pg` for `paragraph`
 * `rails_view` is replaced by `rails_render`. Use `rails_render partial: '/some_partial', locals: { foo: 1 }; rails_render file: '/some_view', locals: { foo: 1 }` to render a partial or a file or anything you want. Use the same as you would use rails `render`
 
-* Syntax to register components has slightly changed. Now you need to call `Matestack::Ui::Component.register(...)` instead of `Matestack::Ui::Core::Component::Registry.register_components`
+* `Matestack::Ui::Core::Component::Registry.register_components` Registry was removed. Write a module with methods to use your components via methods
 
 * an app now defines the complete layout, including the head, scripts, csrf tags etc.